Extract exex processing to node actor crate
