+++ crates/defi-actors/src/lib.rs
@@ -11,7 +11,7 @@ pub use market::{
 pub use market_state::{preload_market_state, MarketStatePreloadedOneShotActor};
 pub use mempool::MempoolActor;
 pub use mergers::{ArbSwapPathMergerActor, DiffPathMergerActor, SamePathMergerActor};
-pub use node::{NodeBlockActor, NodeBlockActorConfig, NodeMempoolActor};
+pub use node::{loom_exex, mempool_worker, NodeBlockActor, NodeBlockActorConfig, NodeMempoolActor};
 pub use node_exex_grpc::NodeExExGrpcActor;
 pub use node_player::NodeBlockPlayerActor;
 pub use pathencoder::SwapEncoderActor;
