+++ crates/defi-actors/src/node/mod.rs
@@ -1,5 +1,6 @@
 pub use node_block_actor::{NodeBlockActor, NodeBlockActorConfig};
 pub use node_mempool_actor::NodeMempoolActor;
+pub use reth_exex_worker::{loom_exex, mempool_worker};
 
 mod node_block_actor;
 mod node_block_hash_worker;
@@ -7,4 +8,5 @@ mod node_block_logs_worker;
 mod node_block_state_worker;
 mod node_block_with_tx_worker;
 mod node_mempool_actor;
+mod reth_exex_worker;
 mod reth_worker;
